### 2021-08-11 #3602 @whedon
[PRE REVIEW]: latentcor: An R Package for estimating latent correlations from mixed data types
**Submitting author:** @muellsen (<a href="http://orcid.org/0000-0002-3821-7083">Christian Lorenz Müller</a>)
**Repository:** <a href="https://github.com/mingzehuang/latentcor" target ="_blank">https://github.com/mingzehuang/latentcor</a>
**Version:** v1.0.0
