### 2021-08-11 4e71230
Add DOIs to references in paper
