+++ paper/paper.bib
@@ -6,7 +6,8 @@
   number={3},
   pages={609--625},
   year={2020},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  doi = {10.1093/biomet/asaa007}
 }
 @article{croux2013robust,
   title={Robust sparse principal component analysis},
@@ -16,7 +17,8 @@
   number={2},
   pages={202--214},
   year={2013},
-  publisher={Taylor \& Francis}
+  publisher={Taylor \& Francis},
+  doi = {10.1080/00401706.2012.727746}
 }
 @article{team2013r,
   title={R: A language and environment for statistical computing},
@@ -32,7 +34,8 @@
   number={2},
   pages={405--421},
   year={2017},
-  publisher={Wiley-Blackwell}
+  publisher={Wiley-Blackwell},
+  doi = {10.1111/rssb.12168}
 }
 @article{liu2009nonparanormal,
   title={The nonparanormal: Semiparametric estimation of high dimensional undirected graphs.},
@@ -54,7 +57,8 @@
   journal={Journal of Computational and Graphical Statistics},
   pages={1--8},
   year={2021},
-  publisher={Taylor \& Francis}
+  publisher={Taylor \& Francis},
+  doi = {10.1080/10618600.2021.1882468}
 }
 
 @manual{fox2019poly,
@@ -72,7 +76,8 @@ title = {polycor: Polychoric and Polyserial Correlations},
   volume={5},
   number={51},
   pages={2306},
-  year={2020}
+  year={2020},
+  doi = {10.21105/joss.02306}
 }
 
 @article{yoon2019microbial,
@@ -82,5 +87,6 @@ title = {polycor: Polychoric and Polyserial Correlations},
   volume={10},
   pages={516},
   year={2019},
-  publisher={Frontiers}
+  publisher={Frontiers},
+  doi = {10.3389/fgene.2019.00516}
 }
