### 2021-08-12 864e077
A more conpact code for interpolation generation.
