### 2021-08-13 dacbf0d
A more compact way.
