### 2021-08-13 25aa610
This will be a wrapper for chebpol so that user doesn't need to generate function value separately.
