+++ R/interpolation.R
@@ -1,18 +1,20 @@
-#Library packages you need for internal functions
-library(foreach)
-library(parallel)
-library(doFuture)
-library(stats)
-library(pcaPP)
-library(fMultivar)
-library(mnormt)
-library(chebpol)
-library(Matrix)
-library(utils)
-library(MASS)
-source('~/latentcor_git/latentcor/R/internal.R')
-
-value = function(evalfun, grid_list, cores = detectCores(), ...) {
+#' @title Parallel version of multilinear interpolation generator for function approximation
+#' @param evalfun The objective function to be approximated.
+#' @param grid_list A list for grid points (each element of list is a vector represents ticklabels on a dimension). The number of list elements are the dimension of function inputs.
+#' @param cores The numbers of cores (threads) of your machine to conduct parallel computing.
+#' @param ... Other inputs for objective functions to be passed through.
+#' @return \code{interpolation} returns
+#' \itemize{
+#'       \item{value: }{A list of of length p corresponding to each variable. Returns NA for continuous variable; proportion of zeros for binary/truncated variables; the cumulative proportions of zeros and ones (e.g. first value is proportion of zeros, second value is proportion of zeros and ones) for ternary variable. }
+#'       \item{interpolant: }{An interpolant function generated by \code{chebpol::chebpol} for interplation.}
+#' @export
+#' @import doFuture
+#' @import foreach
+#' @importFrom parallel detectCores
+#' @importFrom chebpol ipol
+#' @examples
+#'
+interpolation = function(evalfun, grid_list, cores = detectCores(), ...) {
   grid_all = expand.grid(grid_list)
   registerDoFuture()
   plan(multicore, workers = cores)
@@ -26,40 +28,7 @@ value = function(evalfun, grid_list, cores = detectCores(), ...) {
   for (i in 1:d_grid) {
     dim_value = c(dim_value, length(grid_list[[i]]))
   }
-  return (array(as.integer(10^7 * value_vector), dim = dim_value))
+  value = array(value_vector, dim = dim_value)
+  interpolant = ipol(val = value, grid = grid_list, method = "multilin")
+  return (list(value = value, interpolant = interpolant))
 }
-
-evalfun = function(grid_input, comb, tol, ratio) {
-  if (comb == "10" | comb == "20") {
-    zratio1 = grid_input[2]; zratio2 = NA
-  } else if (comb == "11" | comb == "21" | comb == "22") {
-    zratio1 = grid_input[2]; zratio2 = grid_input[3]
-  } else if (comb == "30") {
-    zratio1 = c(grid_input[2] * grid_input[3], grid_input[3]); zratio2 = NA
-  } else if (comb == "31" | comb == "32") {
-    zratio1 = c(grid_input[2] * grid_input[3], grid_input[3]); zratio2 = grid_input[4]
-  } else if (comb == "33") {
-    zratio1 = c(grid_input[2] * grid_input[3], grid_input[3]); zratio2 = grid_input[4:5]
-  }
-  K = grid_input[1] * bound_switch(comb = comb, zratio1 = zratio1, zratio2 = zratio2)
-  r_switch(method = "original", K = K, zratio1 = zratio1, zratio2 = zratio2, comb = comb, tol = tol, ratio = ratio)
[…]
