### 2021-08-13 f3c960b
export interpolation function.
