### 2021-08-13 9e48a5f
Put interpolant generation into folder "supplement"
