+++ R/interpolation.R
@@ -1,25 +1,30 @@
 #' @title Parallel version of multilinear interpolation generator for function approximation
+#' @description Parallel version of multilinear interpolation generator for function approximation
 #' @param evalfun The objective function to be approximated.
 #' @param grid_list A list for grid points (each element of list is a vector represents ticklabels on a dimension). The number of list elements are the dimension of function inputs.
 #' @param cores The numbers of cores (threads) of your machine to conduct parallel computing.
 #' @param ... Other inputs for objective functions to be passed through.
 #' @return \code{interpolation} returns
 #' \itemize{
-#'       \item{value: }{A list of of length p corresponding to each variable. Returns NA for continuous variable; proportion of zeros for binary/truncated variables; the cumulative proportions of zeros and ones (e.g. first value is proportion of zeros, second value is proportion of zeros and ones) for ternary variable. }
+#'       \item{value: }{A list of of length p corresponding to each variable. Returns NA for continuous variable; proportion of zeros for binary/truncated variables; the cumulative proportions of zeros and ones (e.g. first value is proportion of zeros, second value is proportion of zeros and ones) for ternary variable.}
 #'       \item{interpolant: }{An interpolant function generated by \code{chebpol::chebpol} for interplation.}
-#' @export
+#' }
 #' @import doFuture
 #' @import foreach
+#' @import future
 #' @importFrom parallel detectCores
 #' @importFrom chebpol ipol
+#' @export
 #' @examples
-#' interpolation(evalfun = latentcor:::evalfun, grid_list = get(paste("grid_list", comb, sep = "_")), comb = "10", tol = 1e-8, ratio = .9)
-#'
+#' grid_list_10 = list(seq(-0.9, 0.9, by = 0.1), seq(0.1, 0.9, by = 0.1))
+#' interpolation(evalfun = latentcor:::evalfun, grid_list = grid_list_10,
+#'               comb = "10", tol = 1e-8, ratio = .9)
 
 interpolation = function(evalfun, grid_list, cores = detectCores(), ...) {
   grid_all = expand.grid(grid_list)
   registerDoFuture()
   plan(multicore, workers = cores)
+  j = NULL
   value_vector =
     foreach (j = 1:nrow(grid_all), .combine = c) %dopar% {
       grid_input = as.numeric(grid_all[j, ])
@@ -34,3 +39,4 @@ interpolation = function(evalfun, grid_list, cores = detectCores(), ...) {
   interpolant = ipol(val = value, grid = grid_list, method = "multilin")
   return (list(value = value, interpolant = interpolant))
 }
+
