+++ R/internal.R
@@ -130,6 +130,8 @@ zratios = function(X, types) {
   return(out)
 }
 
+combs = c("10", "11", "20", "21", "22", "30", "31", "32", "33")
+
 r_sol = function(K, zratio1, zratio2, comb, tol, ratio) {
   bridge_switch = switch(comb, "10" = function(r, zratio1, zratio2){
     # binary and continuous
@@ -374,9 +376,3 @@ grid_list_31 = list(round(seq(-0.95, 0.95, by = 0.05)), round(seq(0.05, 0.95, by
 grid_list_32 = list(round(seq(-0.95, 0.95, by = 0.05)), round(seq(0.05, 0.95, by = 0.05)), round(seq(0.05, 0.95, by = 0.05)), round(seq(0.05, 0.95, by = 0.05)))
 grid_list_33 = list(seq(-0.9, 0.9, by = 0.1), seq(0.1, 0.9, by = 0.1), seq(0.1, 0.9, by = 0.1), seq(0.1, 0.9, by = 0.1), seq(0.1, 0.9, by = 0.1))
 
-
-combs = c("10", "11", "20", "21", "22", "30", "31", "32", "33")
-for (comb in combs) {
-  assign(paste("ipol", comb, sep = "_"), interpolation(evalfun = evalfun, grid_list = get(paste("grid_list", comb, sep = "_")), comb = comb, tol = 1e-8, ratio = .9))
-}
-save(list = c(paste("ipol", combs, sep = "_")), file = "interpolation.rda", compress = "xz")
