+++ NAMESPACE
@@ -2,6 +2,10 @@
 
 export(GenData)
 export(estR)
+export(interpolation)
+import(doFuture)
+import(foreach)
+import(future)
 import(ggplot2)
 importFrom(MASS,mvrnorm)
 importFrom(Matrix,nearPD)
@@ -10,6 +14,7 @@ importFrom(fMultivar,pnorm2d)
 importFrom(graphics,hist)
 importFrom(heatmaply,heatmaply)
 importFrom(mnormt,pmnorm)
+importFrom(parallel,detectCores)
 importFrom(pcaPP,cor.fk)
 importFrom(plotly,layout)
 importFrom(plotly,plot_ly)
