### 2021-08-13 0b596d7
Add unbias check and numerical evaluation into package.
