### 2021-08-13 c87df36
Add simple example for interpolation function.
