+++ man/interpolation.Rd
@@ -26,7 +26,9 @@ interpolation(evalfun, grid_list, cores = detectCores(), ...)
 Parallel version of multilinear interpolation generator for function approximation
 }
 \examples{
-grid_list_10 = list(seq(-0.9, 0.9, by = 0.1), seq(0.1, 0.9, by = 0.1))
-interpolation(evalfun = latentcor:::evalfun, grid_list = grid_list_10,
-              comb = "10", tol = 1e-8, ratio = .9)
+grid_list = list(seq(-0.9, 0.9, by = 0.1), seq(0.1, 0.9, by = 0.1))
+objfun = function(x, y) {x^2 + sqrt(y)}
+evalfun = function(X) {objfun(X[1], X[2])}
+value = interpolation(evalfun = evalfun, grid_list = grid_list)$value
+interpolant = interpolation(evalfun = evalfun, grid_list = grid_list)$interpolant
 }
