### 2021-08-14 abf92ac
Turn numerical evaluation into general use function.
