### 2021-08-14 dc887d6
debuging evaluation function.
