+++ R/evaluation.R
@@ -1,6 +1,33 @@
+#' @title Numerical evaluation for different estimation methods.
+#' @description Speed and accuracy comparison of two different estimation methods.
+#' @param genfun A data generation function.
+#' @param estfun_1 A function for first estimation method.
+#' @param estfun_2 A function for second estimation method.
+#' @param grid_list A list for grid points to be evaluated (each element of list is a vector represents ticklabels on a dimension). The number of list elements are the dimension of function inputs.
+#' @param nrep Number of replications in simulation.
+#' @param cores The numbers of cores (threads) of your machine to conduct parallel computing.
+#' @param ... Other inputs for data generation or estimation functions to be passed through.
+#' @return \code{evaluation} returns
+#' \itemize{
+#'      \item{meanAE_1: }{An array for mean absolute error of first estimation method.}
+#'      \item{meanAE_2: }{An array for mean absolute error of second estimation method.}
+#'      \item{medianAE_1: }{An array for median absolute error of first estimation method.}
+#'      \item{medianAE_2: }{An array for median absolute error of second estimation method.}
+#'      \item{maxAE_1: }{An array for maximum absolute error of first estimation method.}
+#'      \item{maxAE_2: }{An array for maximum absolute error of second estimation method.}
+#'      \item{meanAE_diff: }{An array for mean absolute error of difference between two estimations.}
+#'      \item{medianAE_diff: }{An array for median absolute error of difference between two estimations.}
+#'      \item{maxAE_diff: }{An array for maximum absolute error of difference between two estimations.}
+#'      \item{mediantime_1: }{An array for median time of first estimation method.}
+#'      \item{mediantime_2: }{An array for median time of second estimation method.}
+#' }
+#' @importFrom stats median
+#' @importFrom microbenchmark microbenchmark
+#' @export
+#' @examples
+#' ## Some example here
 
-
-evaluation = function(genfun, estfun_1, estf_2, grid_list, nrep = 100, cores = detectCores(), ...) {
+evaluation = function(genfun, estfun_1, estfun_2, grid_list, nrep = 100, cores = detectCores(), ...) {
   grid_all = expand.grid(grid_list)
   registerDoFuture()
   plan(multicore, workers = cores)
@@ -8,15 +35,17 @@ evaluation = function(genfun, estfun_1, estf_2, grid_list, nrep = 100, cores = d
   value_vector =
     foreach (j = 1:nrow(grid_all), .combine = c) %dopar% {
       grid_input = as.numeric(grid_all[j, ])
-      time_reps_1 = rep(NA, nrep); time_reps_2 = rep(NA, nrep); error_reps_1 = rep(NA, nrep); error_reps_2 = rep(NA, nrep); diff_reps = rep(NA, nrep)
+      time_reps_1 = time_reps_2 = estimate_reps_1 = estimate_reps_2 = diff_reps = rep(NA, nrep)
       for (r in 1:nrep) {
+        estimate_1 = estimate_2 = NULL
         simdata = genfun(grid_input, ...)
-        time_reps_1[r] = median(microbenchmark::microbenchmark(estimate_reps_1[r] = estfun_1(simdata, ...), times = 5)$time)
[…]
