+++ NAMESPACE
@@ -2,6 +2,7 @@
 
 export(GenData)
 export(estR)
+export(evaluation)
 export(interpolation)
 import(doFuture)
 import(foreach)
@@ -13,11 +14,13 @@ importFrom(chebpol,ipol)
 importFrom(fMultivar,pnorm2d)
 importFrom(graphics,hist)
 importFrom(heatmaply,heatmaply)
+importFrom(microbenchmark,microbenchmark)
 importFrom(mnormt,pmnorm)
 importFrom(parallel,detectCores)
 importFrom(pcaPP,cor.fk)
 importFrom(plotly,layout)
 importFrom(plotly,plot_ly)
+importFrom(stats,median)
 importFrom(stats,na.omit)
 importFrom(stats,optimize)
 importFrom(stats,qnorm)

[…]
