### 2021-08-14 252fbae
Turn numerical evaluation into compact way.
