+++ supplement/num_eval.R
@@ -14,3 +14,142 @@ source("/scratch/user/sharkmanhmz/latentcor_git/latentcor/R/internal.R")
 source("/scratch/user/sharkmanhmz/latentcor_git/latentcor/R/GenData.R")
 source("/scratch/user/sharkmanhmz/latentcor_git/latentcor/R/estR.R")
 source("/scratch/user/sharkmanhmz/latentcor_git/latentcor/R/evaluation.R")
+
+
+latentR = seq(-0.9, 0.9, by = 0.1); zratios = seq(0.1, 0.9, by = 0.1)
+grid_list = list(latentR, zratios)
+
+## For BC case
+genfun = function (rho, zratio) {
+  out = GenData(types = c("bin", "con"), rhos = rho, XP = list(zratio, NA))$X
+  return(out)
+}
+estfun_1 = function(X) {
+  out = estR(X = X, types = c("bin", "con"), method = "original")$R[1, 2]
+  return(out)
+}
+estfun_2 = function(X) {
+  out = estR(X = X, types = c("bin", "con"))$R[1, 2]
+  return(out)
+}
+evaluation_BC = evaluation (genfun = genfun, estfun_1 = estfun_1, estfun_2 = estfun_2, grid_list = grid_list)
+
+## For BB case
+genfun = function (rho, zratio) {
+  out = GenData(types = c("bin", "bin"), rhos = rho, XP = list(zratio, .5))$X
+  return(out)
+}
+estfun_1 = function(X) {
+  out = estR(X = X, types = c("bin", "bin"), method = "original")$R[1, 2]
+  return(out)
+}
+estfun_2 = function(X) {
+  out = estR(X = X, types = c("bin", "bin"))$R[1, 2]
+  return(out)
+}
+evaluation_BB = evaluation (genfun = genfun, estfun_1 = estfun_1, estfun_2 = estfun_2, grid_list = grid_list)
+
+## For TC case
+genfun = function (rho, zratio) {
+  out = GenData(types = c("tru", "con"), rhos = rho, XP = list(zratio, NA))$X
+  return(out)
+}
+estfun_1 = function(X) {
+  out = estR(X = X, types = c("tru", "con"), method = "original")$R[1, 2]
+  return(out)
+}
+estfun_2 = function(X) {
+  out = estR(X = X, types = c("tru", "con"))$R[1, 2]
+  return(out)
+}
+evaluation_TC = evaluation (genfun = genfun, estfun_1 = estfun_1, estfun_2 = estfun_2, grid_list = grid_list)
+
+## For TB case
+genfun = function (rho, zratio) {
+  out = GenData(types = c("tru", "bin"), rhos = rho, XP = list(zratio, 0.5))$X
+  return(out)
+}
+estfun_1 = function(X) {
+  out = estR(X = X, types = c("tru", "bin"), method = "original")$R[1, 2]
+  return(out)
+}
+estfun_2 = function(X) {
+  out = estR(X = X, types = c("tru", "bin"))$R[1, 2]
+  return(out)
+}
+evaluation_TB = evaluation (genfun = genfun, estfun_1 = estfun_1, estfun_2 = estfun_2, grid_list = grid_list)
+
+## For TT case
+genfun = function (rho, zratio) {
+  out = GenData(types = c("tru", "tru"), rhos = rho, XP = list(zratio, 0.5))$X
+  return(out)
+}
+estfun_1 = function(X) {
+  out = estR(X = X, types = c("tru", "tru"), method = "original")$R[1, 2]
+  return(out)
+}
+estfun_2 = function(X) {
+  out = estR(X = X, types = c("tru", "tru"))$R[1, 2]
+  return(out)
+}
[…]
