### 2021-08-15 62e0f9b
Add plot for evaluation function.
