+++ man/evaluation.Rd
@@ -11,7 +11,8 @@ evaluation(
   grid_list,
   nrep = 100,
   cores = detectCores(),
-  showplot = FALSE...
+  showplot = FALSE,
+  ...
 )
 }
 \arguments{
@@ -26,6 +27,8 @@ evaluation(
 \item{nrep}{Number of replications in simulation.}
 
 \item{cores}{The numbers of cores (threads) of your machine to conduct parallel computing.}
+
+\item{showplot}{Logical indicator. \code{showplot = TRUE} generates the heatmaps of output arrays. NULL if \code{showplot = FALSE}.}
 
 \item{...}{Other inputs for data generation or estimation functions to be passed through.}
 }
