+++ supplement/interpolant.R
@@ -1,6 +1,6 @@
 ## The following script generate interpolant on my HPRC. Users should modify source directory and cores to replicate it.
 source("/scratch/user/sharkmanhmz/latentcor_git/latentcor/R/internal.R")
-source("/scratch/user/sharkmanhmz/latentcor_git/latentcor/supplement/interpolant.R")
+source("/scratch/user/sharkmanhmz/latentcor_git/latentcor/R/interpolation.R")
 
 library(MASS)
 library(microbenchmark)
@@ -37,7 +37,7 @@ grid_list_31 = list(round(pnorm(seq(-1.8, 1.8, by =.15), sd = .8), 6) * 2 - 1, r
 grid_list_32 = list(round(pnorm(seq(-1.8, 1.8, by =.15), sd = .8), 6) * 2 - 1, round(pnorm(seq(-1.8, 1.8, by =.15), sd = .8), 6), round(pnorm(seq(-1.8, 1.8, by =.15), sd = .8), 6), round(pnorm(seq(.1, 2.5, by = .1)), 6) * 2 - 1)
 grid_list_33 = list(round(pnorm(seq(-1.8, 1.8, by =.15), sd = .8), 6) * 2 - 1, round(pnorm(seq(-1.8, 1.8, by =.3), sd = .8), 6), round(pnorm(seq(-1.8, 1.8, by =.3), sd = .8), 6), round(pnorm(seq(-1.8, 1.8, by =.3), sd = .8), 6), round(pnorm(seq(-1.8, 1.8, by =.3), sd = .8), 6))
 
-combs = c("10", "11", "20", "21", "22", "30", "31", "32", "33")
+combs = c("10", "11", "20", "21", "22", "30")
 
 for (comb in combs) {
   assign(paste("ipol", comb, sep = "_"), interpolation(evalfun = evalfun, grid_list = get(paste("grid_list", comb, sep = "_")), cores = 80, comb = comb, tol = 1e-8, ratio = .9)$interpolant)
