+++ supplement/interpolant.R
@@ -41,7 +41,10 @@ grid_list_33 = list(round(pnorm(seq(-1.8, 1.8, by =.15), sd = .8), 6) * 2 - 1, r
 combs = c("10", "11", "20", "21", "22", "30", "31", "32", "33")
 
 for (comb in combs) {
-  assign(paste("ipol", comb, sep = "_"), interpolation(evalfun = evalfun, grid_list = get(paste("grid_list", comb, sep = "_")), cores = 80, comb = comb, tol = 1e-8, ratio = .9)$interpolant)
+  ouput = interpolation(evalfun = evalfun, grid_list = get(paste("grid_list", comb, sep = "_")), cores = 80, comb = comb, tol = 1e-8, ratio = .9)
+  assign(paste("ipol", comb, sep = "_"), output$interpolant)
+  assign(paste("value", comb, sep = "_"), output$value)
 }
 save(list = c(paste("ipol", combs, sep = "_")), file = "interpolant.rda", compress = "xz")
+save(list = c(paste("value", combs, sep = "_")), file = "value.rda", compress = "xz")
 #usethis::use_data(ipol_10, ipol_11, ipol_20, ipol_21, ipol_22, ipol_30, ipol_31, ipol_32, ipol_33, internal = TRUE, compress = "xz")
