### 2021-08-16 80ba1e0
Add necessary library for HPRC
