+++ supplement/num_eval.R
@@ -1,13 +1,25 @@
 ## The following script generate interpolant on my HPRC. Users should modify source directory and cores to replicate it.
 
+library(devtools)
+library(usethis)
+library(stats)
+library(fMultivar)
+library(mnormt)
+library(utils)
+library(parallel)
+library(future)
 library(MASS)
-library(microbenchmark)
 library(foreach)
 library(Matrix)
 library(chebpol)
 library(pcaPP)
 library(doRNG)
 library(doFuture)
+library(heatmaply)
+library(ggplot2)
+library(plotly)
+library(graphics)
+library(microbenchmark)
 
 load("/scratch/user/sharkmanhmz/latentcor_git/latentcor/R/sysdata.rda")
 source("/scratch/user/sharkmanhmz/latentcor_git/latentcor/R/internal.R")
@@ -20,137 +32,137 @@ latentR = seq(-0.9, 0.9, by = 0.1); zratios = seq(0.1, 0.9, by = 0.1)
 grid_list = list(latentR, zratios)
 
 ## For BC case
-genfun = function (rho, zratio) {
-  out = GenData(types = c("bin", "con"), rhos = rho, XP = list(zratio, NA))$X
+genfun = function (grid_input) {
+  out = GenData(rhos = grid_input[1], XP = list(grid_input[2], NA))$X
   return(out)
 }
 estfun_1 = function(X) {
-  out = estR(X = X, types = c("bin", "con"), method = "original")$R[1, 2]
+  out = estR(X = X, method = "original")$R[1, 2]
   return(out)
 }
 estfun_2 = function(X) {
-  out = estR(X = X, types = c("bin", "con"))$R[1, 2]
+  out = estR(X = X)$R[1, 2]
   return(out)
 }
-evaluation_BC = evaluation(genfun = genfun, estfun_1 = estfun_1, estfun_2 = estfun_2, grid_list = grid_list, showplot = TRUE)
+evaluation_BC = evaluation(genfun = genfun, estfun_1 = estfun_1, estfun_2 = estfun_2, grid_list = grid_list, showplot = TRUE, types = c("bin", "con"))
 
 ## For BB case
-genfun = function (rho, zratio) {
-  out = GenData(types = c("bin", "bin"), rhos = rho, XP = list(zratio, .5))$X
+genfun = function (grid_input) {
+  out = GenData(rhos = grid_input[1], XP = list(grid_input[2], .5))$X
   return(out)
 }
 estfun_1 = function(X) {
-  out = estR(X = X, types = c("bin", "bin"), method = "original")$R[1, 2]
+  out = estR(X = X, method = "original")$R[1, 2]
   return(out)
 }
 estfun_2 = function(X) {
-  out = estR(X = X, types = c("bin", "bin"))$R[1, 2]
+  out = estR(X = X)$R[1, 2]
   return(out)
 }
-evaluation_BB = evaluation(genfun = genfun, estfun_1 = estfun_1, estfun_2 = estfun_2, grid_list = grid_list, showplot = TRUE)
+evaluation_BB = evaluation(genfun = genfun, estfun_1 = estfun_1, estfun_2 = estfun_2, grid_list = grid_list, showplot = TRUE, types = c("bin", "bin"))
 
 ## For TC case
-genfun = function (rho, zratio) {
-  out = GenData(types = c("tru", "con"), rhos = rho, XP = list(zratio, NA))$X
+genfun = function (grid_input) {
+  out = GenData(rhos = grid_input[1], XP = list(grid_input[2], NA))$X
   return(out)
 }
 estfun_1 = function(X) {
[…]
