### 2021-08-17 9d89f30
specify cores for HPRC.
