### 2021-08-17 03f7bb5
Add doRNG
