+++ NAMESPACE
@@ -5,6 +5,7 @@ export(estR)
 export(evaluation)
 export(interpolation)
 import(doFuture)
+import(doRNG)
 import(foreach)
 import(future)
 import(ggplot2)
