### 2021-08-18 #3602 @crvernon
@mingzehuang @irinagain  @muellsen The JOSS editorial board reviewed your submission and found that it does meet our standards to proceed to review.  I'll be taking this on as your topic editor.
