### 2021-08-18 #3602 @muellsen
@crvernon This is amazing

> @mingzehuang @irinagain @muellsen The JOSS editorial board reviewed your submission and found that it does meet our standards to proceed to review. I'll be taking this on as your topic editor.

@crvernon Thanks so much for these great news! We'll discussing tomorrow potential reviewers and get back to you!
