+++ NAMESPACE
@@ -2,12 +2,6 @@
 
 export(GenData)
 export(estR)
-export(evaluation)
-export(interpolation)
-import(doFuture)
-import(doRNG)
-import(foreach)
-import(future)
 import(ggplot2)
 importFrom(MASS,mvrnorm)
 importFrom(Matrix,nearPD)
@@ -15,13 +9,10 @@ importFrom(chebpol,ipol)
 importFrom(fMultivar,pnorm2d)
 importFrom(graphics,hist)
 importFrom(heatmaply,heatmaply)
-importFrom(microbenchmark,microbenchmark)
 importFrom(mnormt,pmnorm)
-importFrom(parallel,detectCores)
 importFrom(pcaPP,cor.fk)
 importFrom(plotly,layout)
 importFrom(plotly,plot_ly)
-importFrom(stats,median)
 importFrom(stats,na.omit)
 importFrom(stats,optimize)
 importFrom(stats,qnorm)
