### 2021-08-19 #3602 @rmflight
Sure @crvernon, I think I can review this, as long as I'm not trying to completely match the theoretical explanations in the package vignette to the code.

What kind of timeframe are we looking for a first review?
