### 2021-08-19 #3602 @corybrunson
@crvernon i would be glad to! I can get through preliminaries today and complete initial issues within 3 weeks.
