+++ DESCRIPTION
@@ -24,7 +24,7 @@ Authors@R: c(
            email = "irinag@stat.tamu.edu",
            comment = c(ORCID = "0000-0002-4116-0268")))
 Maintainer: Mingze Huang <mingzehuang@gmail.com>
-Description: The R package *latentcor* provides estimation for latent correlation with mixed data types (continuous, binary, truncated and ternary).
+Description: Estimation of latent correlations with mixed data types:continuous, binary, truncated (or zero-inflated) and ternary. The estimation is based on latent copula Gaussian models.
 Depends: R (>= 3.0.0)
 Imports: stats, rmarkdown, pcaPP, fMultivar, mnormt, chebpol, Matrix, utils, MASS, heatmaply, ggplot2, plotly, graphics, foreach, parallel, doFuture, future, microbenchmark, doRNG
 License: GPL-3 + file LICENSE
