### 2021-08-19 #3602 @crvernon
Alright @corybrunson @rmflight @muellsen  I am going to close this Pre-Review and kick off the full review which you should receive a notification for.  Thanks!
