### 2021-08-19 #3634 @crvernon
Also, please don't forget to add a link to this review issue in any issues or pull requests you may generate in the https://github.com/mingzehuang/latentcor repository.  This will help everyone have a single point of reference.
