### 2021-08-25 #3634 @crvernon
:mega:  Mid-week rally!  Just checking in to see how things are going.

:open_hands: @corybrunson @rmflight I am happy to answer any questions you may have!

Have a great rest of the week!
