### 2021-08-26 ed50576
expand estR documentation on tol, nu and ratio
