+++ R/latentcor.R
@@ -1,7 +1,7 @@
 #' @title Estimate latent correlation for mixed types.
 #' @description Estimation of latent correlation matrix from observed data of (possibly) mixed types (continuous/binary/truncated/ternary) based on the latent Gaussian copula model. Missing values (NA) are allowed. The estimation is based on pairwise complete observations.
-#' @rdname estR
-#' @aliases estR
+#' @rdname latentcor
+#' @aliases latentcor
 #' @param X A numeric data matrix (n by p), where n is number of samples, and p is number of variables. Missing values (NA) are allowed, in which case the estimation is based on pairwise complete observations.
 #' @param types A vector of length p indicating the type of each of the p variables in \code{X}. Each element must be one of \code{"con"} (continuous), \code{"bin"} (binary), \code{"tru"} (truncated) or \code{"ter"} (ternary). If the vector has length 1, then all p variables are assumed to be of the same type that is supplied. The default value is \code{"con"} which means all variables are continuous.
 #' @param method The calculation method for latent correlations. Either \code{"original"} or \code{"approx"}. If \code{method = "approx"}, multilinear approximation method is used, which is much faster than the original method, see Yoon et al. (2021). If \code{method = "original"}, optimization of the bridge inverse function is used. The default is \code{"approx"}.
@@ -9,7 +9,7 @@
 #' @param tol When \code{method = "original"}, specifies the desired accuracy of the bridge function inversion via uniroot optimization and is passed to \code{\link{optimize}}. The default value is 1e-8. When \code{method = "approx"}, this parameter is ignored.
 #' @param ratio When \code{method = "approx"}, specifies the boundary value for multilinear interpolation, must be between 0 and 1. When \code{ratio = 0}, no linear interpolation is performed (the slowest execution) which is equivalent to \code{method = "original"}. When \code{ratio = 1}, linear interpolation is always performed (the fastest execution) but may lead to high approximation errors. The default (recommended) value is 0.9 which controls the approximation error and has fast execution, see Yoon et al. (2021) for details. When \code{method = "original"}, this parameter is ignored.
 #' @param showplot Logical indicator. \code{showplot = TRUE} generates a ggplot object \code{plotR} with the heatmap of latent correlation matrix \code{R}. \code{plotR = NULL} if \code{showplot = FALSE}.
-#' @return \code{estR} returns
+#' @return \code{latentcor} returns
 #' \itemize{
 #'       \item{zratios: }{A list of of length p corresponding to each variable. Returns NA for continuous variable; proportion of zeros for binary/truncated variables; the cumulative proportions of zeros and ones (e.g. first value is proportion of zeros, second value is proportion of zeros and ones) for ternary variable. }
 #'       \item{K: }{(p x p) Kendall Tau (Tau-a) Matrix for \code{X} }
@@ -34,9 +34,9 @@
 #' @importFrom pcaPP cor.fk
 #' @importFrom chebpol ipol
 #' @export
-#' @example man/examples/estR_ex.R
+#' @example man/examples/latentcor_ex.R
 
-estR = function(X, types = "con", method = c("approx", "original"), nu = 0.001, tol = 1e-8, ratio = 0.9, showplot = FALSE){
+latentcor = function(X, types = "con", method = c("approx", "original"), nu = 0.001, tol = 1e-8, ratio = 0.9, showplot = FALSE){
   if(nu < 0 | nu > 1){
     stop("nu must be be between 0 and 1.")
   } else if(tol <= 0) {
