### 2021-08-27 adc9696
Update function name to latentcor
