+++ README.md
@@ -41,13 +41,13 @@ library(latentcor)
 X = GenData(types = c("ter", "con"), XP = list(c(0.3, .5), NA))$X
 
 # Estimate latent correlation matrix with the original method
-estR(X = X, types = c("ter", "con"), method = "original")$R
+latentcor(X = X, types = c("ter", "con"), method = "original")$R
 
 # Estimate latent correlation matrix with the approximation method
-estR(X = X, types = c("ter", "con"))$R
+latentcor(X = X, types = c("ter", "con"))$R
 
 # Heatmap for latent correlation matrix.
-estR(X = X, types = c("ter", "con"), showplot = TRUE)$plotR
+latentcor(X = X, types = c("ter", "con"), showplot = TRUE)$plotR
 ```
 Another example with the `mtcars` dataset.
 
@@ -58,13 +58,13 @@ X = mtcars
 # Check variable types
 apply(mtcars, 2, table)
 # Estimate latent correlation matrix with original method
-estR(mtcars, types = c("con", "ter", "con", "con", "con", "con", "con", "bin",
+latentcor(mtcars, types = c("con", "ter", "con", "con", "con", "con", "con", "bin",
                        "bin", "ter", "con"), method = "original")$R
 # Estimate latent correlation matrix with approximation method
-estR(mtcars, types = c("con", "ter", "con", "con", "con", "con", "con", "bin",
+latentcor(mtcars, types = c("con", "ter", "con", "con", "con", "con", "con", "bin",
                        "bin", "ter", "con"))$R
 # Heatmap for latent correlation matrix with approximation method.
-estR(mtcars, types = c("con", "ter", "con", "con", "con", "con", "con", "bin",
+latentcor(mtcars, types = c("con", "ter", "con", "con", "con", "con", "con", "bin",
                        "bin", "ter", "con"), showplot = TRUE)$plotR
 ```
 
