+++ vignettes/latentcor.Rmd
@@ -59,13 +59,13 @@ head(X, n = 6L)
 simdata$plotX
 ```
 
-Then we can estimate the latent correlation matrix based on these 2 variables using `estR` function.
+Then we can estimate the latent correlation matrix based on these 2 variables using `latentcor` function.
 
 ```{r estimation}
-estimate = estR(X, types = c("ter", "con"))
+estimate = latentcor(X, types = c("ter", "con"))
 ```
 
-The output of `estR` is a list with several elements:
+The output of `latentcor` is a list with several elements:
 
 ```{r estimation_output}
 names(estimate)
@@ -95,7 +95,7 @@ estimate$Rpointwise
 estimate$R
 ``` 
 
-  - `plotR`: NULL by default as `showplot = FALSE` in `estR`. Otherwise displays a heatmap of latent correlation matrix.
+  - `plotR`: NULL by default as `showplot = FALSE` in `latentcor`. Otherwise displays a heatmap of latent correlation matrix.
 
 ```{r heatmap}
 estimate$plotR
@@ -115,13 +115,13 @@ Let's take a look at the unique values for each variable to determine the corres
 apply(mtcars, 2, table)
 ```
 
-Then we can estimate the latent correlation matrix for all variables of `mtcars` by using `estR` function.
+Then we can estimate the latent correlation matrix for all variables of `mtcars` by using `latentcor` function.
 
 ```{r mtcars_estimation}
-estimate_mtcars = estR(mtcars, types = c("con", "ter", "con", "con", "con", "con", "con", "bin", "bin", "ter", "con"))
+estimate_mtcars = latentcor(mtcars, types = c("con", "ter", "con", "con", "con", "con", "con", "bin", "bin", "ter", "con"))
 ```
 
-The output of `estR` for `mtcars`:
+The output of `latentcor` for `mtcars`:
 
 ```{r mtcars_estimation_output}
 names(estimate_mtcars)
@@ -151,7 +151,7 @@ estimate_mtcars$Rpointwise
 estimate_mtcars$R
 ``` 
 
-  - `plotR`: NULL by default as `showplot = FALSE` in `estR`. Otherwise displays a heatmap of latent correlation matrix for `mtcars` (See [heatmap of latent correlation (approx) for mtcars](https://rpubs.com/mingzehuang/797937)).
+  - `plotR`: NULL by default as `showplot = FALSE` in `latentcor`. Otherwise displays a heatmap of latent correlation matrix for `mtcars` (See [heatmap of latent correlation (approx) for mtcars](https://rpubs.com/mingzehuang/797937)).
 
 ```{r mtcars_heatmap}
 estimate_mtcars$plotR
@@ -228,7 +228,7 @@ where $n$ is the sample size.
 `latentcor` calculates pairwise Kendall's $\widehat \tau$ as part of the estimation process
 
 ```{r KendallTau}
-estimate = estR(X, types = c("con", "bin", "ter", "tru"))
+estimate = latentcor(X, types = c("con", "bin", "ter", "tru"))
 K = estimate$K
 K
 ```
@@ -361,7 +361,7 @@ Given the form of bridge function $F$, obtaining a moment-based estimation $\wid
 Original estimation approach relies on numerical inversion of $F$ based on solving uni-root optimization problem.
 
 ```{r callR, warning = FALSE, message = F}
-estimate = estR(X, types = c("con", "bin", "ter", "tru"), method = "original")
+estimate = latentcor(X, types = c("con", "bin", "ter", "tru"), method = "original")
 ```
 
 ***Algorithm for Original method***
@@ -404,7 +404,7 @@ In short, d-dimensional multi-linear interpolation uses a weighted average of $2
 
 
[…]

[…]
