### 2021-08-27 mingzehuang/latentcor#8 @rmflight
paper: should it be interpretation of linear relatedness?
Starting on [Line 28 of the PDF of the paper draft](https://raw.githubusercontent.com/openjournals/joss-papers/joss.03634/joss.03634/10.21105.joss.03634.pdf), I see this statement that seems to need some clarification:

> the resulting values do not have correlation interpretation and can not be used as direct substitutes in statistical methods that require correlation as input

Are you stating that they don't have a **linear correlation** interpretation, and therefore aren't appropriate for other methods? I would really like to see a reference for this claim, and rewording to make it clearer.
