### 2021-08-27 779faae
Minor vignette adjustments
