### 2021-08-28 mingzehuang/latentcor#9 @rmflight
paper: too much detail about the implementation and API
Per my own understanding, JOSS papers should have a:

[x] summary
[x] statement of need
[ ] almost nothing else

The current draft includes

* Details about the speedy approximation implementation that seem out of place besides mentioning it in passing
* Examples of usage and differences with respect to Pearson that should be in a vignette
