### 2021-08-28 #3634 @rmflight
Issues around the paper itself:

* a line where I think the statement isn't quite correct: mingzehuang/latentcor#8
* The paper includes a lot more than is typical for a JOSS manuscript, including API and examples of running. mingzehuang/latentcor#9
