### 2021-08-28 #3634 @rmflight
The current manuscript shows **some** of the things I'd like to see in another vignette, and shows that the software runs properly, and I've verified it can run, and it has tests, so I think I'm done for a first round.

My verdict is **Major Revisions**. Please feel free to ask for more clarification on the issues if they aren't clear.
