### 2021-08-28 0f7c362
Update function names, finish get_types.
