+++ tests/testthat/test_zratio.R
@@ -2,21 +2,21 @@
 library(latentcor)
 
 test_that("zratio is NULL if type is continuous", {
-  X = GenData(types = "con", copulas = "no")$X
+  X = gen_data(types = "con", copulas = "no")$X
   expect_true(is.na(unlist(zratios(X = X, types = "con"))))
 })
 
 test_that("zratio is the proportion of zeros if type is binary", {
-  X = GenData(types = "bin", copulas = "no")$X
+  X = gen_data(types = "bin", copulas = "no")$X
   expect_equal(unlist(zratios(X = X, types = "bin")), as.numeric(colMeans(as.matrix(X == 0))))
 })
 
 test_that("zratio is the proportion of zeros if type is truncated", {
-  X = GenData(types = "tru", copulas = "no")$X
+  X = gen_data(types = "tru", copulas = "no")$X
   expect_equal(unlist(zratios(X = X, types = "tru")), as.numeric(colMeans(as.matrix(X == 0))))
 })
 
 test_that("zratio is the proportion of zeros and ones if type is ternary", {
-  X = GenData(types = "ter", copulas = "no")$X
+  X = gen_data(types = "ter", copulas = "no")$X
   expect_equal(unlist(zratios(X = X, types = "ter")), as.numeric(c(colMeans(as.matrix(X == 0)), 1 - colMeans(as.matrix(X == 2)))))
 })
