+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: latentcor
 Type: Package
 Title: Fast Computation of Latent Correlations for Mixed Data
-Version: 1.0.0.9000
+Version: 1.0.1
 Authors@R: c(
     person(given = "Mingze",
            family = "Huang",
