### 2021-08-28 579412f
Add NEWS.md
