+++ man/gen_data.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/GenData.R
+% Please edit documentation in R/gen_data.R
 \name{gen_data}
 \alias{gen_data}
 \title{Mixed type simulation data generator}
