### 2021-08-31 134527a
Make a pre_process function to deal with data types and non-numeric cases.
