### 2021-08-31 7fb3a00
make pre_process function and create an example for it.
