+++ NAMESPACE
@@ -2,6 +2,7 @@
 
 export(gen_data)
 export(latentcor)
+export(pre_process)
 import(ggplot2)
 importFrom(MASS,mvrnorm)
 importFrom(Matrix,nearPD)
