+++ R/pre_process.R
@@ -13,15 +13,19 @@
 #' raw_data$X1 = 1:3
 #' raw_data$X2 = c("medium", "small", "large")
 #' raw_data$X3 = 7:9
-#' pre_process = pre_process(raw_data = raw_data, ordinals = list(NA, c("small", "medium", "large")))
+#' pre_process = pre_process(raw_data = raw_data,
+#'               ordinals = list(NA, c("small", "medium", "large"), NA))
 #' pre_process$X
 #' pre_process$types
 
 pre_process = function(raw_data, ordinals = NULL) {
-  X = data.frame(raw_data); n = nrow(X); p = ncol(X); types = rep(NA, p)
+  raw_data = data.frame(raw_data); n = nrow(raw_data); p = ncol(raw_data); types = rep(NA, p)
+  X = matrix(NA, n, p)
   for (i in 1:p) {
-    if (!(is.numeric(X[ , i]))) {
-      X[ , i] = ord(X[ , i], ordinals[[i]])
+    if (is.numeric(raw_data[ , i])) {
+      X[ , i] = raw_data[ , i]
+    } else {
+      X[ , i] = ord(raw_data[ , i], ordinals[[i]])
     }
     level = unique(X[ , i])
     if (length(level) <= 1) {
@@ -43,5 +47,6 @@ pre_process = function(raw_data, ordinals = NULL) {
       }
     }
   }
-  return(list(X = data.matrix(X), types = types))
+  colnames(X) = colnames(raw_data)
+  return(list(X = X, types = types))
 }
