+++ man/pre_process.Rd
@@ -26,7 +26,8 @@ raw_data = data.frame(matrix(NA, 3, 3))
 raw_data$X1 = 1:3
 raw_data$X2 = c("medium", "small", "large")
 raw_data$X3 = 7:9
-pre_process = pre_process(raw_data = raw_data, ordinals = list(NA, c("small", "medium", "large")))
+pre_process = pre_process(raw_data = raw_data,
+              ordinals = list(NA, c("small", "medium", "large"), NA))
 pre_process$X
 pre_process$types
 }
