+++ NEWS.md
@@ -1,3 +1,0 @@
-# latentcor 1.1.0
-
-* Added a `NEWS.md` file to track changes to the package.
