### 2021-08-31 4dc428b
Adding details to latentcor documentation
