### 2021-08-31 9cafea4
Expanded details in documentation for latentcor concerning parameters ratio and nu
