### 2021-09-01 9e87a85
Write changelog for v1.1.0.
