+++ NEWS.md
@@ -1,3 +1,6 @@
 # latentcor 1.1.0
 
-* Added a `NEWS.md` file to track changes to the package.
+* Rename main function from `estR` to `latentcor`.
+* Add help function `get_types` to automatically generate types for data matrix.
+* Add more documentation on parameters.
+* Add more documentation for speed comparison between original method and approximation.
