### 2021-09-01 8eda853
Minor document clarification on get_types
