### 2021-09-01 5a2edcb
Parameter clarifications in vignette
