### 2021-09-01 1c84491
Add documentation for default type NULL.
