### 2021-09-01 3a92525
Correct the bug in latentcor if types = NULL
