### 2021-09-01 1cfce38
Adjust logic of automatic truncated type assignment to only focus on zero-inflation
