### 2021-09-01 eb05769
Add example of using get_types to vignette
