+++ vignettes/latentcor.Rmd
@@ -115,11 +115,18 @@ Let's take a look at the unique values for each variable to determine the corres
 apply(mtcars, 2, table)
 ```
 
+
 Then we can estimate the latent correlation matrix for all variables of `mtcars` by using `latentcor` function.
 
 ```{r mtcars_estimation}
 estimate_mtcars = latentcor(mtcars, types = c("con", "ter", "con", "con", "con", "con", "con", "bin", "bin", "ter", "con"))
 ```
+
+Note that the determination of variable types can also be done automatically by `latentcor` package using `get_types` function:
+```{r mtcars_types, message = FALSE}
+get_types(mtcars)
+```
+This function is run automatically inside `latentcor` if the `types` are not supplied by the user, however the automatic determination of types takes extra time, so we recommend to specify `types` explicitly if they are known in advance.
 
 The output of `latentcor` for `mtcars`:
 
