### 2021-09-01 ec3674b
Add illustration of get_types to readme
