### 2021-09-01 cfdeeaa
Add timing example to latentcor documentation
