### 2021-09-02 059d09b
Add speed comparison in README.md
