### 2021-09-02 #3634 @crvernon
:mega: Hello all!  It looks like @rmflight  and @corybrunson have made good progress on the review!  Also, great work @mingzehuang and @irinagain in addressing their issues as they arise.

**Just a reminder for the reviewers:**  don't forget to check off the reviewer checklist as you feel your comments have been sufficiently addressed.  Also, once you believe your review to be completed, shoot me a comment here.

👏 Great progress everyone and keep up the good work!  Let's keep driving towards getting this software publication ready!
