### 2021-09-02 1934069
Add timing to compare original method and different level of approximation.
