### 2021-09-02 15bc3f2
Add subsampling example to vignette
