### 2021-09-02 mingzehuang/latentcor#10 @corybrunson
functionality: specify performance claims
The software performs very much as advertised, with even greater improvements in performance on larger data sets that i've tried (generated using the helpful `gen_data()` function)! However, this performance improvement is not quantified in the documentation, so far as i've found. Presumably this is contained in the cited papers and perhaps also in the software paper. Because performance is the core motivation for the package, though, i think it important to also summarize quantified performance gains (whether theoretical or experimental) in the documentation, e.g. the README. (This would also allow me, as a reviewer, to evaluate the claims.)

_Part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/3634)._
