### 2021-09-02 mingzehuang/latentcor#11 @corybrunson
documentation: statement of need
Similarly to #10, the software documentation should include a clear statement of need. This could be a single sentence summarizing motivations from the software paper, located in the DESCRIPTION, the README, and/or the manual page for `latentcor()`.

_Part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/3634).
