### 2021-09-02 mingzehuang/latentcor#12 @corybrunson
split vignettte?
The vignette is an extremely helpful part of the installation, though it runs quite long and seems to naturally divide into halves: practical use and theoretical underpinnings. Would it make sense to split it into separate vignettes that users can read based on need? For example, based on the first few screens' worth, i wouldn't guess that the vignette also included mathematical formulations and would assume that i would have to consult the papers. (This is a question, not a request, and is not holding up my review!)

_Part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/3634)._
