### 2021-09-02 mingzehuang/latentcor#13 @corybrunson
documentation: recognized input classes
This is a small point, but one that might cause confusion for users new to R. The documentation explains that the input `X` to `latentcor()` should be a numeric data matrix, but this could mean either an object of class 'matrix' or of class 'data.frame'. Both are understood by `latentcor()` though this is not always the case, so i would suggest making explicit that either data type can be passed, or that the input will be coerced to a numeric matrix using `data.matrix()`.

_Part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/3634)._
