### 2021-09-02 mingzehuang/latentcor#14 @corybrunson
paper: clarity of examples
In a similar spirit to #9, i'm more open to more than bare-bones content in the software paper, but i do think the paper should be structured differently than documentation—for readability rather than for reference. In particular, the two examples are explained not in exposition within the paper but in comments embedded with the code. This makes them hard to follow on their own and to compare (e.g. syntactically) to each other. The heat map of differences between the method is useful, for instance, but i think it would be easier to interpret with a short paragraph of text than with the code chunks used to generate it.

_Part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/3634)._
