### 2021-09-02 #3634 @corybrunson
The package is quite easy to use and i fully intend to make use of it myself, so i thank the authors for creating it! I have reviewed the package and the paper and raised all issues that occurred to me. I may raise more as others are addressed, but for the time being i welcome responses (whether revisions or feedback) from the authors. : )
