### 2021-09-03 c886d5f
Add statement of need to README.md
