### 2021-09-03 21842de
Add statement of need in README.md
