+++ README.md
@@ -21,16 +21,15 @@
 
 ## Statement of need
 No R software package is currently available that allows accurate and fast correlation estimation from mixed variable data in a unifying manner. 
-The popular `cor` function within R package `stats` [@team2013r], for instance, allows to compute Pearson's correlation, Kendall's $\tau$ and Spearman's
-$\rho$, and a faster algorithm for calculating Kendall's $\tau$ is implemented in the R package `pcaPP` [@croux2013robust]. Pearson's correlation is not
+The popular [`cor`](https://www.rdocumentation.org/packages/stats/versions/3.6.2/topics/cor) function within R package [`stats`](https://rdrr.io/r/stats/stats-package.html), for instance, allows to compute Pearson's correlation, Kendall's $\tau$ and Spearman's
+$\rho$, and a faster algorithm for calculating Kendall's $\tau$ is implemented in the R package [`pcaPP`](https://cran.r-project.org/web/packages/pcaPP/index.html). Pearson's correlation is not
 appropriate for skewed or ordinal data, and its use leads to invalid inference in those cases. While the rank-based Kendall's $\tau$ and Spearman's $\rho$ are
-more robust measures of association, they cannot directly be used as subsitutes for statistical methods that require Pearson correlation as input (e.g., graphical model estimation [@yoon2019microbial]). The R package `polycor` [@fox2019poly] is designed for ordinal data and allows to computes
+more robust measures of association, they cannot directly be used as subsitutes for statistical methods that require Pearson correlation as input (e.g., graphical model estimation [Yoon et al. 2019](https://www.frontiersin.org/articles/10.3389/fgene.2019.00516/full?utm_source=S-TWT&utm_medium=SNET&utm_campaign=ECO_FGENE_XXXXXXXX_auto-dlvrit)). The R package [`polycor`](https://cran.r-project.org/web/packages/polycor/index.html) is designed for ordinal data and allows to computes
 polychoric (ordinal/ordinal) and polyserial (ordinal/continuous) correlations based on latent Gaussian model. However, the package does not have functionality
-for zero-inflated data, nor can it handle skewed continuous measurements as it does not allow for copula transformation. The R package `correlation`
-[@makowski2020methods] in the `easystats` collection provides 16 different correlation measures, including polychoric and polyserial correlations. However, 
-functionality for correlation estimation from zero-inflated data is lacking. The R package `mixedCCA` [@yoon2020sparse] is based on the latent Gaussian copula 
+for zero-inflated data, nor can it handle skewed continuous measurements as it does not allow for copula transformation. The R package [`correlation`](https://cran.r-project.org/web/packages/correlation/index.html) in the [`easystats`](https://cran.r-project.org/web/packages/see/index.html) collection provides 16 different correlation measures, including polychoric and polyserial correlations. However, 
+functionality for correlation estimation from zero-inflated data is lacking. The R package [`mixedCCA`](https://cran.r-project.org/web/packages/mixedCCA/index.html) is based on the latent Gaussian copula 
 model and can compute latent correlations between continuous/binary/zero-inflated variable types as an intermediate step for canonical correlation analysis. 
-However, `mixedCCA` does not allow for ordinal data types. The R package `latentcor`, introduced here, thus represents the first stand-alone R package for 
+However, [`mixedCCA`](https://cran.r-project.org/web/packages/mixedCCA/index.html) does not allow for ordinal data types. The R package [`latentcor`](https://cran.r-project.org/web/packages/latentcor/index.html), introduced here, thus represents the first stand-alone R package for 
 computation of latent correlation that takes into account all variable types (continuous/binary/ordinal/zero-inflated), comes with an optimized memory footprint, 
 and is computationally efficient, essentially making latent correlation estimation almost as fast as rank-based correlation estimation. 
 
