+++ vignettes/latentcor.bib
@@ -80,3 +80,25 @@ title = {polycor: Polychoric and Polyserial Correlations},
   note = {R package version 2.1-2},
   url = {https://github.com/sgaure/chebpol},
 }
+
+@article{makowski2020methods,
+  title={Methods and algorithms for correlation analysis in R},
+  author={Makowski, Dominique and Ben-Shachar, Mattan S and Patil, Indrajeet and L{\"u}decke, Daniel},
+  journal={Journal of Open Source Software},
+  volume={5},
+  number={51},
+  pages={2306},
+  year={2020},
+  doi = {10.21105/joss.02306}
+}
+
+@article{yoon2019microbial,
+  title={Microbial networks in SPRING-Semi-parametric rank-based correlation and partial correlation estimation for quantitative microbiome data},
+  author={Yoon, Grace and Gaynanova, Irina and M{\"u}ller, Christian L},
+  journal={Frontiers in genetics},
+  volume={10},
+  pages={516},
+  year={2019},
+  publisher={Frontiers},
+  doi = {10.3389/fgene.2019.00516}
+}
