### 2021-09-03 mingzehuang/latentcor#11 @mingzehuang
Thank you for your suggestion, Dr.Brunson! We've added statement of need into README.md and Vignette. Also a sentence for statement of need in DESCRIPTION.
Let us know any more needed:)
