### 2021-09-03 mingzehuang/latentcor#11 @corybrunson
@mingzehuang they do the job! To be clear, the requirement for the documentation is not the same as for the paper, so it would be fine if there were merely a sentence or two summary rather than a full section if that's what you'd prefer. But the issue is certainly resolved.
