### 2021-09-03 b5e9db6
Both numeric matrix and data frame are OK
