### 2021-09-03 22d33e4
check data type rather than coerce into numeric directly.
