### 2021-09-03 mingzehuang/latentcor#13 @mingzehuang
Thank you for your detailed review, @corybrunson ! The input data can be either matrix or data frame as long as it's numeric. We've modify documentation to declare it explicitly. It should be checked if it's numeric rather than simply coerced. We've modified the code for it! Thank you so much!
