### 2021-09-03 mingzehuang/latentcor#8 @muellsen
Thanks a lot for the comment. To avoid any confusion about semantics or interpretation, we tried to make our statement more precise. The sentence reads now

While the rank-based Kendall’s τ and Spearman’s ρ are more robust measures of _association_,
they cannot directly be used as subsitutes for statistical methods that require Pearson correlation as input (a prominent example is, e.g., graphical model estimation (Xue & Zou, 2012;Yoon et al., 2019)).

Here, we also added the Xue & Zou reference. The paper succinctly describes the difficulty of using rank-based estimators in the context of graphical model estimation and provides a solution to it for continuous data.

More generally, rank-based estimators are not mere "generalizations" of Pearson correlation but, indeed, measure a different "type" of relationship (see, e.g., the nice example in Wikipedia's correlation article https://en.wikipedia.org/wiki/Correlation in the Subsection on Rank correlations (and references therein)). This is what we originally meant by "correlation interpretation" (but this was obviously too vague).
