### 2021-09-03 mingzehuang/latentcor#9 @muellsen
Thanks for the comment! We agree that some JOSS papers do have just this basic structure but there are certainly many other examples, e.g., [here](https://www.theoj.org/joss-papers/joss.02306/10.21105.joss.02306.pdf), a paper that is similar in spirit, but structured differently, a recent [contribution](https://joss.theoj.org/papers/10.21105/joss.02844) of my colleagues and myself, or this [one](https://www.theoj.org/joss-papers/joss.03393/10.21105.joss.03393.pdf). For this contribution here, we would like to follow a similar route. Also, if I understand @corybrunson's [comment](https://github.com/mingzehuang/latentcor/issues/14) correctly, we may embed code and text a bit better by expanding the mtcars example (and connect the ternary example better to the code).

In fact, one question here is whether it would be OK to add a speed-up plot into the paper? This would bolster the performance claims (in addition to the reduced memory footprint).
