### 2021-09-03 mingzehuang/latentcor#10 @muellsen
We already started to add microbenchmarks/ timings to the vignette/README. In your view, would it be also beneficial to have a speed-up plot in the paper?
