+++ vignettes/latentcor.Rmd
@@ -209,522 +209,6 @@ for (s in 1:10){
 proc.time() - start_time
 ```
 
-# Mathematical framework for estimation
-
-## Latent Gaussian Copula Model for Mixed Data
-
-`latentcor` utilizes the powerful semi-parametric latent Gaussian copula models to estimate latent correlations between mixed data types (continuous/binary/ternary/truncated or zero-inflated). Below we review the definitions for each type.
-
-***Definition of continuous model*** [@fan2017high]
-
-A random $X\in\cal{R}^{p}$ satisfies the Gaussian copula (or nonparanormal) model if there exist monotonically increasing $f=(f_{j})_{j=1}^{p}$ with $Z_{j}=f_{j}(X_{j})$ satisfying $Z\sim N_{p}(0, \Sigma)$, $\sigma_{jj}=1$; we denote $X\sim NPN(0, \Sigma, f)$.
-
-```{r continuous}
-X = gen_data(n = 6, types = "con")$X
-X
-```
-
-***Definition of binary model*** [@fan2017high]
-
-A random $X\in\cal{R}^{p}$ satisfies the binary latent Gaussian copula model if there exists $W\sim NPN(0, \Sigma, f)$ such that $X_{j}=I(W_{j}>c_{j})$, where $I(\cdot)$ is the indicator function and $c_{j}$ are constants.
-
-```{r binary}
-X = gen_data(n = 6, types = "bin")$X
-X
-```
-
-***Definition of ternary model*** [@quan2018rank]
-
-A random $X\in\cal{R}^{p}$ satisfies the ternary latent Gaussian copula model if there exists $W\sim NPN(0, \Sigma, f)$ such that $X_{j}=I(W_{j}>c_{j})+I(W_{j}>c'_{j})$, where $I(\cdot)$ is the indicator function and $c_{j}<c'_{j}$ are constants.
-
-```{r ternary}
-X = gen_data(n = 6, types = "ter")$X
-X
-```
-
-***Definition of truncated or zero-inflated model*** [@yoon2020sparse]
-
-A random $X\in\cal{R}^{p}$ satisfies the truncated latent Gaussian copula model if there exists $W\sim NPN(0, \Sigma, f)$ such that $X_{j}=I(W_{j}>c_{j})W_{j}$, where $I(\cdot)$ is the indicator function and $c_{j}$ are constants.
-
-```{r truncated}
-X = gen_data(n = 6, types = "tru")$X
-X
-```
-
-
-***Mixed latent Gaussian copula model***
-
-The mixed latent Gaussian copula model jointly models $W=(W_{1}, W_{2}, W_{3}, W_{4})\sim NPN(0, \Sigma, f)$ such that $X_{1j}=W_{1j}$, $X_{2j}=I(W_{2j}>c_{2j})$, $X_{3j}=I(W_{3j}>c_{3j})+I(W_{3j}>c'_{3j})$  and $X_{4j}=I(W_{4j}>c_{4j})W_{4j}$.
-
-```{r mixed}
-set.seed("234820")
-X = gen_data(n = 100, types = c("con", "bin", "ter", "tru"))$X
-head(X)
-```
-
-## Moment-based estimation of $\Sigma$ based on bridge functions
-
-
[…]
