+++ vignettes/latentcor_math.Rmd
@@ -1,0 +1,544 @@
+---
+title: "Mathematical Framework for latentcor"
+author: "Mingze Huang, Christian L. Müller, Irina Gaynanova"
+date: "`r Sys.Date()`"
+bibliography: latentcor.bib
+output: rmarkdown::html_vignette
+extra_dependencies: ["amsmath"]
+nocite: |
+  @croux2013robust
+  @filzmoser2021pcapp
+  @liu2009nonparanormal
+  @fox2019poly
+vignette: >
+  %\VignetteIndexEntry{Mathematical Framework for latentcor}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+```{r setup, include = FALSE}
+options(tinytex.verbose = TRUE)
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>"
+)
+library(latentcor)
+```
+
+# Latent Gaussian Copula Model for Mixed Data
+
+`latentcor` utilizes the powerful semi-parametric latent Gaussian copula models to estimate latent correlations between mixed data types (continuous/binary/ternary/truncated or zero-inflated). Below we review the definitions for each type.
+
+***Definition of continuous model*** [@fan2017high]
+
+A random $X\in\cal{R}^{p}$ satisfies the Gaussian copula (or nonparanormal) model if there exist monotonically increasing $f=(f_{j})_{j=1}^{p}$ with $Z_{j}=f_{j}(X_{j})$ satisfying $Z\sim N_{p}(0, \Sigma)$, $\sigma_{jj}=1$; we denote $X\sim NPN(0, \Sigma, f)$.
+
+```{r continuous}
+X = gen_data(n = 6, types = "con")$X
+X
+```
+
+***Definition of binary model*** [@fan2017high]
+
+A random $X\in\cal{R}^{p}$ satisfies the binary latent Gaussian copula model if there exists $W\sim NPN(0, \Sigma, f)$ such that $X_{j}=I(W_{j}>c_{j})$, where $I(\cdot)$ is the indicator function and $c_{j}$ are constants.
+
+```{r binary}
+X = gen_data(n = 6, types = "bin")$X
+X
+```
+
+***Definition of ternary model*** [@quan2018rank]
+
+A random $X\in\cal{R}^{p}$ satisfies the ternary latent Gaussian copula model if there exists $W\sim NPN(0, \Sigma, f)$ such that $X_{j}=I(W_{j}>c_{j})+I(W_{j}>c'_{j})$, where $I(\cdot)$ is the indicator function and $c_{j}<c'_{j}$ are constants.
+
+```{r ternary}
+X = gen_data(n = 6, types = "ter")$X
+X
+```
+
+***Definition of truncated or zero-inflated model*** [@yoon2020sparse]
+
+A random $X\in\cal{R}^{p}$ satisfies the truncated latent Gaussian copula model if there exists $W\sim NPN(0, \Sigma, f)$ such that $X_{j}=I(W_{j}>c_{j})W_{j}$, where $I(\cdot)$ is the indicator function and $c_{j}$ are constants.
+
+```{r truncated}
+X = gen_data(n = 6, types = "tru")$X
+X
+```
+
+
[…]
