### 2021-09-03 mingzehuang/latentcor#12 @mingzehuang
Thank you @corybrunson ! We've split the vignette into two parts: one called "latentcor", another called "latentcor_math". So that users can choose whatever they want to read.
Let us know if anything could be improved:)
