### 2021-09-03 mingzehuang/latentcor#10 @corybrunson
I think that would suffice! I'll have a look at the vignette again when they're included.

Since your claims involve memory as well as time, it would be good to include comments on both.

In addition to example runs, if there are more comprehensive results in the cited journal articles, then i think it would also be important to mention them.
