### 2021-09-03 mingzehuang/latentcor#10 @mingzehuang
@corybrunson See the microbenchmark for both examples in README.md [here](https://mingzehuang.github.io/latentcor/index.html). Also microbenchmark in the code chunk in section `Approximation method (method = "approx")` in Vignette ['latentcor_math'](https://mingzehuang.github.io/latentcor/articles/latentcor_math.html).
Thank you very much!
