### 2021-09-09 19d0629
Rewrite example to have in text rather than comments
