### 2021-09-09 mingzehuang/latentcor#14 @irinagain
We have rewritten the example to remove comments as much as possible and explain things in text.
