### 2021-09-09 mingzehuang/latentcor#14 @corybrunson
Great, thanks! I will have a look and close the issue after the next PDF update at the review.
