### 2021-09-09 mingzehuang/latentcor#14 @corybrunson
Hm, i see what you have done and i do think it improves the readability. Though it introduces a new problem: code chunks are included but it's not clear what their outputs look like. (I'm not sure, but i think the previous version included returned values from the function calls, as in a vignette.)

Based on other software papers, i can think of a couple of styles that might read better as an article:
1. For each of the two examples, include a single code chunk with all the essential steps (with or without outputs), and refer to the lines in that chunk as needed in the text.
2. Omit code chunks entirely and just mention inline what functions or options are being used, e..g "Applying the original estimation method, using the parameter setting `method = "original"`, gives an estimated latent correlation of ...".

Do those make sense? I do think either would work.
