### 2021-09-10 cfe1a6f
Adjust examples to be in one code chunk
