### 2021-09-10 mingzehuang/latentcor#14 @irinagain
We didn't have returned values in the paper before, only in vignette, because the paper has to be in .md (whereas the vignette is in .Rmd). We believe it's useful to have the code chunk for new users (even though they are still present in vignette). We followed the option 1 above where we consolidated code chunks back into a single one for each example, and have an extended paragraph beforehand commenting on the chunk. Let us know if this works. Thanks!
