### 2021-09-10 mingzehuang/latentcor#10 @corybrunson
Thank you! I'm sorry, i meant to come back to this days ago. I do think the examples suffice to illustrate the claim.

However, asking again to be sure: Do the cited papers report more systematic justifications for the claims, e.g. simulations over a range of scales or big-O proofs of runtime? If not, then i think it will be fine to stop here. If so, i think it would be good to mention these more rigorous results and cite the papers that report them. Does that make sense?
