### 2021-09-11 mingzehuang/latentcor#10 @corybrunson
I took a look at the approximation paper (Yoon &al, 2021), and i see that it makes a thorough case for improved performance with little cost to accuracy. So for this issue i would be satisfied by a sentence in the README or the performance vignette (or both) explaining that this more comprehensive demonstration of the performance claims can be found in that paper.
