### 2021-09-13 4737f57
Done with speed comparison section in vignette latentcor_math
