### 2021-09-13 mingzehuang/latentcor#10 @mingzehuang
Thank you for your update @corybrunson ! In the math vignette "latentcor_math", we just added a section "Speed Comparison" under "Approximation method (method = "approx")" with performance claim and log-log plot for different dimensions.
Let us know if any more need to be added. Thanks!
