### 2021-09-13 mingzehuang/latentcor#10 @mingzehuang
[latentcor_math](https://mingzehuang.github.io/latentcor/articles/latentcor_math.html)
