### 2021-09-13 mingzehuang/latentcor#10 @corybrunson
@mingzehuang this looks good! I might have missed the references to the 2021 paper on a previous read, but together with the plot the vignette is now very clear about the performance claims.
