### 2021-09-13 #3634 @corybrunson
@crvernon the issues i raised have been addressed and i have completed my review, though i welcome the opportunity to review further revisions while the review remains open.
