### 2021-09-15 9b0f515
Add figure to README.md
