### 2021-09-19 mingzehuang/latentcor#9 @rmflight
You've shown that others have gone this route, that's fine. I'm OK with it.
