### 2021-09-19 #3634 @rmflight
Apologies for taking so long to finish looking at the responses to my issues.

I believe everything I brought up has been addressed. The authors have an awesome package, and I definitely have some more background reading to do on the different correlation measures.
