### 2021-09-20 #3634 @crvernon
@mingzehuang please add the following missing DOI to your citation in the paper:

```
- 10.1214/12-aos1041 may be a valid DOI for title: Regularized rank-based estimation of high-dimensional nonparanormal graphical models
```
