### 2021-09-20 #3634 @crvernon
@mingzehuang - Your paper is looking great - and very clean I might add!  Here are several things I would like you to address from my review of your paper before we move on:

- [ ]  Please check the missing DOI in the references (see the above `whedon check references` output ☝️)
- [ ] Please take one last look at your author names and their affiliations to make sure all is well before we move forward
