+++ paper/paper.bib
@@ -100,5 +100,6 @@ title = {polycor: Polychoric and Polyserial Correlations},
   pages={2541--2571},
   year={2012},
   publisher={Institute of Mathematical Statistics}
+  doi={10.1214/12-aos1041}
 }
 
