+++ paper/paper.bib
@@ -99,7 +99,7 @@ title = {polycor: Polychoric and Polyserial Correlations},
   number={5},
   pages={2541--2571},
   year={2012},
-  publisher={Institute of Mathematical Statistics}
+  publisher={Institute of Mathematical Statistics},
   doi={10.1214/12-aos1041}
 }
 
