+++ vignettes/latentcor.bib
@@ -102,3 +102,16 @@ title = {polycor: Polychoric and Polyserial Correlations},
   publisher={Frontiers},
   doi = {10.3389/fgene.2019.00516}
 }
+
+@article{xue2012regularized,
+  title={Regularized rank-based estimation of high-dimensional nonparanormal graphical models},
+  author={Xue, Lingzhou and Zou, Hui},
+  journal={The Annals of Statistics},
+  volume={40},
+  number={5},
+  pages={2541--2571},
+  year={2012},
+  publisher={Institute of Mathematical Statistics},
+  doi={10.1214/12-aos1041}
+}
+
