### 2021-09-20 #3634 @crvernon
@mingzehuang we are almost there!  Next is just setting up the archive for your new release.  Since there have not been any releases yet in your repo, here are a few pointers:

- I find Zenodo easy to use to mint a DOI (permanent archive identifier) for each GitHub release.  You will likely need to create a free Zenodo account and link your GitHub repository to it.  Then once linked, you can conduct your release on GitHub and it will automatically trigger Zenodo to mint a DOI for you.  If you want the DOI as a badge, you can add that as well.  There are other options for creating a permanent archive resulting in a DOI for your code as well.  Don't hesitate to reach out if you have questions.
- We want to make sure the archival has the correct metadata that JOSS requires.  This includes a title that matches the paper title and a correct author list.

So here is what we have left to do:
- [ ] Conduct a GitHub release of the current reviewed version of the software you now have on the `master` and archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository)
- [ ] Check the archival deposit (e.g., in Zenodo) to ensure it has the correct metadata. This includes the title (should match the paper title) and author list. You may also add the authors' ORCID.
- [ ] Please list the DOI of the archived version here
